fixes for ivan commits

+++ chain.go
@@ -7,14 +7,14 @@ import (
 // ChainAttrs are the attributes of a Chain
 type Chain struct {
 	Parent uint32
-	Chain  uint16
+	Chain  uint32
 }
 
 func (c Chain) String() string {
 	return fmt.Sprintf("{Parent: %d, Chain: %d}", c.Parent, c.Chain)
 }
 
-func NewChain(parent uint32, chain uint16) Chain {
+func NewChain(parent uint32, chain uint32) Chain {
 	return Chain{
 		Parent: parent,
 		Chain:  chain,
